Fix sentient actors slowly sliding down slopes
This was goofy because slope resistance was done at the /end/ of the
frame, for reasons that make sense to me but that I'm having a hard time
putting into words.  Hm.

Anyway, I moved that to the beginning of the frame, and found a new
problem: velocity integration was doing funny things to gravity (as
intended), so slope resistance was overcompensating, which made sentient
actors slowly slide UP slopes.  Argh.

I fixed this by introducing pending_force, a sum of acceleration vectors
that should apply this frame, and integrating over that.  Both gravity
and slope resistance are included in it, so they exactly cancel out
regardless of how the integration works.