[WIP] Vastly improve push physics
This tracks pushing recursively and treats the pusher as the controller
of a set of objects.  Momentum is conserved when a new object is added
to the set; an object leaving the set is given appropriate velocity; and
friction of the whole set applies to the pusher's movement.

Lots of lingering minor bugs and cleanups, but this is the best result
I've had in a week.