Vastly improve handling of slides and overlaps
This eliminates some redundant stuff (like the minimum penetration
vector, which it turns out wasn't necessary) and starts expressing the
behavior of the collision in terms of when it will start and when it
will end.  Overlap handling in particular is a lot more solid now.

Also added a set of "standardized" collision tests that always use
100x100 boxes and check much more of the output, just to be sure that
various forms of simple collision are handled correctly.