[WIP] Final attempt at momentum-based pushing
This almost works, but it's not what I want.  It still allows "ramming"
(throwing yourself at a too-heavy crate over and over, nudging it a
couple pixels at a time), which is a side effect of both semi-realistic
physics and discrete timesteps.  It makes the game feel sloppy and bad
and I don't like it.

Upon further reflection of the kinds of gameplay effects I actually want
in fox flux, I've decided to abandon the realistic approach entirely,
rip out most of this code, and do something more direct and cartoony.

It's a huge bummer to make this call after all the effort I've put into
trying to make this work, so here's the final state of it for posterity.
It has zero momentum — in other words, may it rest in peace.