Slightly remodel push physics and start cleaning it up
The Walk component is no longer responsible for adjusting your movement
speed when pushing; instead, Tote has (finally) figured out how to do
it in a way that prevents ramming!  A lot of Tote references are now
gone from other components, too, and a bunch of this code can still be
nuked.

There are some small remaining issues, but the behavior I get here is
pretty satisfying.  Finally.